fix(telegram): break spool retry import cycle
(cherry picked from commit 4ed8460418e604b1cefbbe825d65daf7cb11ec5d)
